own_trello/requirements.txt
atlassian-python-api==3.41.9
Django==5.0.2
pymongo==4.6.1
requests==2.31.0
urllib3==2.2.0
whitenoise==6.6.0
gunicorn